refactor: simplify web folder paths
