+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "build": "tsc --project tsconfig.build.json && tsc-alias -p tsconfig.build.json",
     "lint": "tsc --noEmit && eslint .",
-    "dev": "tsx watch src/web/api/server | pino-pretty",
+    "dev": "tsx watch src/web/server | pino-pretty",
     "test": "jest",
     "test:unit": "jest --selectProjects unit-tests",
     "test:functional": "jest --selectProjects functional-tests",
