+++ tests/api/client.ts
@@ -1,7 +1,7 @@
 import Fastify from 'fastify';
 import supertest from 'supertest';
 
-import app from '../../src/web/api/app';
+import app from '@web/app';
 
 export async function makeClient() {
   const fastify = Fastify();
